+++ CHANGELOG.md
@@ -1,14 +1,19 @@
+## [0.0.4]
+
+* New: Manage the fling animation velocity.
+* Improve fling animation.
+
 ## [0.0.3]
 
 * NEW: Enable zooming.
 
 ## [0.0.2+2]
 
-Update README.md.
+* Update README.md.
 
 ## [0.0.2+1]
 
-Improve documentation.
+* Improve documentation.
 
 ## [0.0.2]